+++ src/zope/schema/_compat.py
@@ -1,6 +1,7 @@
 import sys
 
 PY3 = sys.version_info[0] >= 3
+PY2 = not PY3
 
 if PY3:  # pragma: no cover